# src/routes/public.ts
import { Router, Request, Response, NextFunction } from 'express';
import { entityService, verificationService, productResponsibilityService } from '../services';
import { publicRateLimiter } from '../middleware';
import { RoleType, SourceType, VerificationStatusType } from '@prisma/client';

const router = Router();

// Apply rate limiting to all public routes
router.use(publicRateLimiter);

/**
 * @route   GET /api/public/entities
 * @desc    Public read-only access to entities
 * @access  Public (rate-limited)
 */
router.get('/entities', async (req: Request, res: Response, next: NextFunction) => {
    try {
        const result = await entityService.list({
            search: req.query.search as string,
            country: req.query.country as string,
            role: req.query.role as RoleType,
            isActive: 'true', // Only active entities in public API
            limit: req.query.limit as string,
            offset: req.query.offset as string,
        });

        // Return simplified public view
        const publicEntities = result.entities.map((entity) => ({
            id: entity.id,
            name: entity.normalizedName,
            country: entity.normalizedCountry,
            city: entity.normalizedCity,
            roles: entity.roles.map((r) => ({
                type: r.roleType,
                marketContext: r.marketContext,
            })),
        }));

        res.json({
            success: true,
            data: publicEntities,
            pagination: {
                total: result.total,
                limit: result.limit,
                offset: result.offset,
                hasMore: result.offset + result.limit < result.total,
            },
        });
    } catch (error) {
        next(error);
    }
});

/**
 * @route   GET /api/public/entities/:id
 * @desc    Public entity detail
 * @access  Public (rate-limited)
 */
router.get('/entities/:id', async (req: Request, res: Response, next: NextFunction) => {
    try {
        const entity = await entityService.getById(req.params.id);
        const latestVerification = await verificationService.getLatestVerificationStatus(entity.id);

        // Return public view with limited information
        res.json({
            success: true,
            data: {
                id: entity.id,
                name: entity.normalizedName,
                address: entity.normalizedAddress,
                city: entity.normalizedCity,
                country: entity.normalizedCountry,
                website: entity.normalizedWebsite,
                phone: entity.normalizedPhone,
                roles: entity.roles.map((r) => ({
                    type: r.roleType,
                    marketContext: r.marketContext,
                    productScope: r.productScope,
                })),
                verificationStatus: latestVerification?.status || 'UNVERIFIED',
                lastUpdated: entity.updatedAt,
            },
        });
    } catch (error) {
        next(error);
    }
});

/**
 * @route   GET /api/public/schema
 * @desc    Data schema documentation
 * @access  Public (rate-limited)
 */
router.get('/schema', (req: Request, res: Response) => {
    res.json({
        success: true,
        data: {
            version: '1.0.0',
            description: 'OpenGPSR API Schema - Public Reference Data for GPSR Entities',
            disclaimer: 'This data is informational only. It does not constitute legal advice or guarantee compliance with GPSR regulations.',
            entities: {
                description: 'Business entities relevant to GPSR (producers, importers, responsible persons)',
                fields: {
                    id: 'Unique identifier (UUID)',
                    name: 'Normalized entity name',
                    address: 'Business address',
                    city: 'City',
                    country: 'ISO 3166-1 alpha-2 country code',
                    website: 'Entity website URL',
                    email: 'Contact email',
                    phone: 'Contact phone',
                    roles: 'Array of GPSR roles',
                    verificationStatus: 'Current verification status (informational only)',
                },
            },
            roleTypes: Object.values(RoleType),
            sourceTypes: Object.values(SourceType),
            verificationStatuses: Object.values(VerificationStatusType),
            rateLimit: {
                windowMs: 900000,
                maxRequests: 100,
                description: 'Public API is rate-limited to prevent abuse',
            },
        },
    });
});

/**
 * @route   GET /api/public/products/:id/resolved
 * @desc    Public "best known truth" for product responsibilities (v2.0)
 * @access  Public (rate-limited)
 * 
 * @query   country - Required: ISO 3166-1 alpha-2 country code
 * @query   at - Optional: Date for historical query (ISO 8601 format, e.g., 2025-11-20)
 * 
 * Returns the highest-confidence responsible entities for each GPSR role
 * with conflict detection, data freshness indicators, and temporal context.
 */
router.get('/products/:id/resolved', async (req: Request, res: Response, next: NextFunction) => {
    try {
        const { id } = req.params;
        const { country, at } = req.query;

        if (!country) {
            return res.status(400).json({
                success: false,
                error: 'Country code is required (e.g., ?country=PL)',
                legalStatus: 'INFORMATIONAL_ONLY',
            });
        }

        // Parse optional 'at' date for historical queries
        let validOnDate: Date | undefined;
        if (at) {
            validOnDate = new Date(at as string);
            if (isNaN(validOnDate.getTime())) {
                return res.status(400).json({
                    success: false,
                    error: 'Invalid date format. Use ISO 8601 (e.g., 2025-11-20)',
                    legalStatus: 'INFORMATIONAL_ONLY',
                });
            }
        }

        const resolved = await productResponsibilityService.getResolved(
            id,
            country as string,
            validOnDate
        );

        // Simplify for public consumption
        const publicResolved = {
            productId: resolved.productId,
            countryCode: resolved.countryCode,
            context: resolved.context,
            responsibilities: Object.fromEntries(
                Object.entries(resolved.responsibilities).map(([role, data]) => [
                    role,
                    {
                        entityName: data.entity?.normalizedName,
                        entityCountry: data.entity?.normalizedCountry,
                        confidence: data.confidence,
                        dataFreshnessDays: data.dataFreshnessDays,
                        hasConflicts: data.hasConflicts,
                    },
                ])
            ),
            conflictCount: resolved.conflictCount,
        };

        res.json({
            success: true,
            legalStatus: 'INFORMATIONAL_ONLY',
            data: publicResolved,
        });
    } catch (error) {
        next(error);
    }
});

/**
 * @route   GET /api/public/stats
 * @desc    Public statistics about the database
 * @access  Public (rate-limited)
 */
router.get('/stats', async (req: Request, res: Response, next: NextFunction) => {
    try {
        const [
            totalEntities,
            activeEntities,
            totalSources,
            entitiesByCountry,
        ] = await Promise.all([
            entityService.list({ limit: '1' }).then((r) => r.total),
            entityService.list({ isActive: 'true', limit: '1' }).then((r) => r.total),
            require('../config/database').default.source.count(),
            require('../config/database').default.entity.groupBy({
                by: ['normalizedCountry'],
                _count: true,
                where: { isActive: true },
                orderBy: { _count: { normalizedCountry: 'desc' } },
                take: 10,
            }),
        ]);

        res.json({
            success: true,
            data: {
                totalEntities,
                activeEntities,
                totalSources,
                topCountries: entitiesByCountry.map((c: any) => ({
                    country: c.normalizedCountry,
                    count: c._count,
                })),
                lastUpdated: new Date().toISOString(),
            },
        });
    } catch (error) {
        next(error);
    }
});

export default router;